Fixes #12925 - Reduced Repository Model size
According to our rubocop metrics, max length per class is 500 lines.
Repository.rb is very close to crossing that.

One approach is to split functionality to different modules and include
it in the main model.

This commit aims to address that by moving content related functionality to
individual modules and having em included in the model file
